Nav rotorcraft rework (#2964)
This is a complete rework of the navigation for rotorcraft with the key features:

- all nav functions and interface in float
- nav function are registered (decoupling between nav API and implementation of standard pattern)
- submodes to specify if setpoint is a position (legacy), speed or accel (better integration of algorithms like GVF)
- guidance (H and V) is reorganized with default function to implement (run_pos, run_speed, run_accel) 
- guidance control (the old default PID) is separated from the guidance logic

---------

+++ sw/airborne/modules/stereocam/nav_line_avoid/avoid_navigation.h
@@ -51,7 +51,7 @@ extern bool obstacle_detected;
 
 void init_avoid_navigation(void);
 void run_avoid_navigation_onvision(void);
-extern void increase_nav_heading(int32_t *heading, int32_t increment);
+extern void increase_nav_heading(float *heading, float increment);
 
 
 #endif /* AVOID_NAVIGATION_H */
